LPD-85584 Move AsyncPicker to common
